chore: 테스트 위치 수정
